builtin pages: restructure archives to facilitate browsing (WIP)
This commit starts the process of merging the 'archives' and 'archive' views into a single view, with an archives listing nav on the left, and the currently-selected archive on the right.

It also starts the process of cleaning up the code around these views, which is (honestly) not very good.