builtin pages: reimplement the history view
This commit makes the archive subnav work, and adds the history view.
There are some minor differences from the old history view, mainly to
improve clarity and simplify.

As an aside, I'm abusing global variables to get around the somewhat
half-baked approach to UI and app-state I'm using. (See the 'rerender' in
the archive-view.js.)