builtin pages: about view for archives, and start of site 'types'
The about view takes a simple tabular format. It *could* be used to display
the readme.md as well, but I'm on the fence about that. Maybe. It would
certainly fit well here.

This commit officially "commits" to the premise of giving dats a "type."
I havent decided how many official types there will be, or how the types
will be set (yet). But I've started with Applications, Users, and Websites
(the last being the generic default).

Some of the dat-types will serve special roles. Users will be used (at
minimum) to publish dats, via the public peer services. Applications will
play a role in the app:// scheme.

This commit also fixes a break in the styling for archive page errors.