Use individual swarms for dats.
Also: use datland-swarm-defaults, and close any new connection that
fails to handshake the archive within a few seconds.

It turns out that discovery-swarm only allows one connection to
occur between peers at once. This leaves us with two options:

 A) Have discovery-swarm emit a special notice when discovery has
    happened but a connection exists, so that we can multiplex in
    the new streams.

 B) Maintain one discovery-swarm per archive.

The dat team pushed me to use option B, as they felt that is
simpler for now. We'll re-evaluate in the future.

Switching to datland-swarm-defaults helps centralize some config
that we dont really want to get out of sync.

It's sometimes possible for a failure to happen on the other peer
that causes the handshake not to occur, but without an error
condition. (This has been the case, in some situations, with
hypercloud.) The connection-timeout code corrects for that, by
closing the connection if the stream fails to start, rather than
letting it stay open as a zombie.