Do not auto-download the content of feeds.
This is part of a rather nuanced decision about how to use Dat.
Chiefly: unless the full history of the archive is retained via
external storage, you cannot prioritize the download of the full
content-feed.

Why? It's, in fact, somewhat obvious if you think about it: if
you aren't intent on storing the full history, and your external
storage doesn't properly give each historical version a unique
name, then you are going to end up writing new and old copies of
entries onto the same file as their pieces arrive.

So we must be much more specific about which blocks we request,
unless we intend to store the history, which we don't yet intend
to do.

The solution is simple enough: only download the files on-demand.
We'll add facilities for downloading *all* files in an archive
at some point later.