Put protocol identifiers, not site/dat titles, in navbar.
This is important for security. Previously, Beaker was placing the
'title' in dat.json in the left part of the URL bar, with green
coloring, similar to how EV SSL certs are rendered in Chrome. This
almost certainly could be used to mess with people. It became
abundantly clear to me when I created a dat with the title
"datprotocol.com". Oh, geez, that's misleading isnt it? Why not
throw in "https://" to really round out my attack.

Yes, this is an obvious mistake. It was motivated by wanting more
useful information in the URL bar-- dat hashes are not very
user-friendly. But, this is obviously not the right solution.

To give better visual cues, I've used a scheme where HTTP gets the
old gray info icon, HTTPS gets the green lock with "Secure", and
DAT gets a blue share-icon with "P2P". That should help people
quickly identify what they're looking at.