Yet another fix to deal with overflowing URLs.
If a URL is very long, it overflows the URL bar. We need the URL bar
to stay within a certain width, rather than expand to accomodate the
URL.

Overflow: hidden is causing the autocomplete dropdown to disappear.

This fix, instead, sets a fixed width relative to the viewport size.
That width will have to be recalculated each time buttons are added
to the navbar. C'est la vie.