Factor out dat-dns and pauls-dat-api modules.
This is a first step in stabilizing the internal workings of Dat
for Beaker. These two modules are better documented and tested
than the internals previously were.

Next step is to redesign and /externalize/ Beaker's APIs such that
userland applications are able to use them. Most (all?) currently
internal APIs will be put behind a permissions barrier, and part
of the next step will be upgrading the permissions system to
accomodate a more diverse system.