Don't show spinner favicon on new tab
Neither Firefox nor Chrome show the spinner icon when opening a new
tab. This makes it feel more like a part of the browser instead of an
external web page.

I do wonder if this makes the new tab load time more noticeable.
If it does, oh well. At least it will give us motivation to shave down
the load time as much as possible :)