Update language used for saving/unsaving sites
This changes the language we use in the sidebar for saving and unsaving
sites.

Owners are presented with a button that toggles the state between
"Delete" and "Restore".

Non-owners see a button that says "Sync", that when clicked opens a
dropdown menu with two items: "Sync for offline use" and "Online only".