Make the modals optional for DatArchive.create() and DatArchive.fork()
Prior to this commit, these calls would automatically create modals
which the user employed to follow through with the action.

Those modals never made anybody very happy. They are slow to initiate
and they rarely fit into the flow of an app making the call.

This commit adds a new option, {prompt: true}, to both calls. That opt
shows the modals as before. It defaults to false.