Add DatArchive class to the background process.
This was an unexpected kludge, required in order to integrate InjestDB into
the background process. (InjestDB consumes the DatArchive API.)

The wrapper calls to the dat-archive web api code. It pretends to be an RPC
call from 'beaker://internal/', and should work fine.

I dislike the bloat of this solution, but it should perform well enough and
be correct.