Add the beaker-profiles-api Injest.
This commit moves Beaker's databases in a new direction by integrating
InjestDB into its core model. This is to support browser features which
public and consume data on the Dat network.

In this commit, we update the profile records and bookmark records to
use the Injest.

The bookmarks implementation is a bit kludgey because Injest is not yet
capable of writing 'private' records. That capability will be added when
Dat's multiwriter is implemented, which will enable private overlays of
the files archive. For now, Beaker integrates two sources of truth: the
private bookmarks sqlite, and the public bookmarks injest.