Cache favicons to the sitedata db and read from cache if a dat isnt available.
Reading the favicon from the dat requires the dat to be loaded and swarmed. In
some cases, this would cause bookmarks not to have a favicon, because the dat
hadn't been loaded yet (eg by a visit).

This change caches the favicons to the sitedata DB, and then falls back to the
sitedata DB cache if the dat isnt available.

It writes to the cache more frequently than it needs to right now. A future
optimization might only write if the data has changed.