Try putting the location input back into the location subwindow.
Prior to this commit, the location subwindow was not being given focus
when it was opened. IF I REMEMBER CORRECTLY(!) this was because of a
performance lag during input which I found disruptive. However, I
BELIEVE(!) that lag was actually caused by window resizing rather than
the initial lag of showing the window. Therefore, it should be safe to
give the location focus again, and to move the input into the location
subwindow.

The *reason* I want to give focus to the location subwindow is that
there are clickable items inside the subwindow, and those items need
to handle hover states.

This commit may need to be reverted after testing on Linux and Windows.